server/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
newsapi-python==0.2.1
nltk==3.2.5
oauthlib==2.0.7
requests==2.13.0
requests-oauthlib==0.8.0
six==1.11.0
textblob==0.15.1
urllib3==1.22
Werkzeug==0.14.1
